# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Antipa API Documentation",
        "description": "Documentation de l'API Antipa",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reportcases": {
            "get": {
                "summary": "Récupérer la liste des cas reportés",
                "operationId": "40d9613a95219c7fd22767510c4f8092",
                "responses": {
                    "200": {
                        "description": "Liste des cas reportés"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage des cas reportés."
                    }
                }
            },
            "post": {
                "summary": "Créer un nouveau cas reporté",
                "operationId": "1c37349c64ac8635c94bc3bfa45565ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "totalConfirmed": {
                                        "type": "integer"
                                    },
                                    "totalDeaths": {
                                        "type": "integer"
                                    },
                                    "totalActive": {
                                        "type": "integer"
                                    },
                                    "dateInfo": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "diseaseId": {
                                        "type": "integer"
                                    },
                                    "localizationId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cas reporté créé avec succès."
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la création du cas reporté."
                    }
                }
            }
        },
        "/api/reportcases/{id}": {
            "get": {
                "summary": "Afficher un cas reporté spécifique",
                "operationId": "9d1a566ff3f7f1173422418a77ee6eda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du cas reporté"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage du cas reporté."
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour un cas reporté spécifique",
                "operationId": "fc90568fccf34500ed5118e106caafd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "totalConfirmed": {
                                        "type": "integer"
                                    },
                                    "totalDeaths": {
                                        "type": "integer"
                                    },
                                    "totalActive": {
                                        "type": "integer"
                                    },
                                    "dateInfo": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "diseaseId": {
                                        "type": "integer"
                                    },
                                    "localizationId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cas reporté mis à jour avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la mise à jour du cas reporté."
                    }
                }
            },
            "delete": {
                "summary": "Supprimer un cas reporté spécifique",
                "operationId": "e26a6ddacb095790d6e23d15919a90c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cas reporté supprimé avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la suppression du cas reporté."
                    }
                }
            }
        }
    }
}